scraper/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
prometheus-client==0.20.0
python-dotenv==1.0.1
